third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/svg-use-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not use cross-origin SVG use content</title>
  <link rel="help" href="https://github.com/WICG/html-in-canvas">
  <link rel="author" href="mailto:schenney@chromium.org">
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <svg id="child-same" x=0 y=0 width=100 height=100>
      <use x=0 y=0  href="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/green-100x100.svg" />
    </svg>
    <svg id="child-cross" x=0 y=0 width=100 height=100>
      <use x=0 y=0 href="https://{{hosts[alt][www]}}:{{ports[h2][0]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.svg" />
    </svg>
  </canvas>

  <script>
    window.onload = () => {
      test(function(t) {
        var context = canvas.getContext("2d", { willReadFrequently: true });
        context.drawElementImage(document.getElementById("child-same"), 0, 0);
        context.drawElementImage(document.getElementById("child-cross"), 0, 100);

        pixel = context.getImageData(50, 50, 1, 1).data;
        assert_equals(pixel[0], 0, "Same origin should draw red");
        assert_equals(pixel[1], 255, "Same origin should draw green");
        assert_equals(pixel[2], 0, "Same origin should draw blue");

        // Cross-origin <use> is not rendered at all.
        pixel = context.getImageData(50, 150, 1, 1).data;
        assert_equals(pixel[0], 0, "Cross origin should not draw red");
        assert_equals(pixel[1], 0, "Cross origin should not draw green");
        assert_equals(pixel[2], 0, "Cross origin should not draw blue");

        t.done();
      });
    }
  </script>
</body>
</html>
